+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Updated the project CI to use Piral Pipelines with YAML
 * Updated `withApi` to use `context` instead of `converters`
 * Updated `piral-jest-utils` package for Jest mocking (#186)
+* Updated Piral CLI path resolution for Windows (#192)
 * Added `piral-update` as a plugin to integrate auto update
 * Added `feed` parameter to `pilet debug` (#344)
 * Added `userStore` as an option for `piral-oidc` (#345)